implement fillDocument method, need to grab the form fields
